move-only: section CoinJoin-specific code to separate source file
Functions are in order of declaration in `wallet/wallet.h`. Review with
`git log -p -n1 --color-moved=dimmed_zebra`
